adding optional kwargs for returning Response object directly
With the `return_response_object` kwarg the reponse object can be
returned right away before any other actions are performed on it (or
partially returned). This is usefull for example if this client is used
as a proxy client and the full HTTP response (so including code) should
be returned.